add Sonar and release plugin. Fix test: remove jmockit improve coverage.

+++ src/main/java/jsr223/docker/compose/DockerComposeScriptEngine.java
@@ -1,12 +1,9 @@
 package jsr223.docker.compose;
 
 import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Reader;
 import java.io.StringWriter;
-import java.io.Writer;
-import java.nio.file.FileAlreadyExistsException;
 import java.util.Map;
 
 import javax.script.AbstractScriptEngine;
@@ -16,20 +13,27 @@ import javax.script.ScriptEngineFactory;
 import javax.script.ScriptException;
 import javax.script.SimpleBindings;
 
-import jsr223.docker.compose.script.environment.EnvironmentVariablesAdder;
+import jsr223.docker.compose.bindings.MapBindingsAdder;
+import jsr223.docker.compose.bindings.StringBindingsAdder;
+import jsr223.docker.compose.file.write.ConfigurationFileWriter;
 import jsr223.docker.compose.utils.DockerComposePropertyLoader;
+import jsr223.docker.compose.utils.DockerComposeVersionGetter;
+import jsr223.docker.compose.yaml.VariablesReplacer;
 import lombok.extern.log4j.Log4j;
-import org.jetbrains.annotations.NotNull;
 import processbuilder.SingletonProcessBuilderFactory;
 import processbuilder.utils.ProcessBuilderUtilities;
 
 @Log4j
 public class DockerComposeScriptEngine extends AbstractScriptEngine {
 
-    private static final EnvironmentVariablesAdder environmentVariablesAdder = new EnvironmentVariablesAdder();
-
     private static final String DOCKER_HOST_PROPERTY_NAME = "DOCKER_HOST";
-    private static final String log4jConfigurationFile = "config/log/scriptengines.properties";
+    private static final String LOG4J_CONFIGURATION_FILE = "config/log/scriptengines.properties";
+
+    private static ProcessBuilderUtilities processBuilderUtilities = new ProcessBuilderUtilities();
+    private static VariablesReplacer variablesReplacer = new VariablesReplacer();
+    private static ConfigurationFileWriter configurationFileWriter = new ConfigurationFileWriter();
+    private static StringBindingsAdder stringBindingsAdder = new StringBindingsAdder(
+            new MapBindingsAdder());
 
 
     public DockerComposeScriptEngine() {
@@ -38,12 +42,12 @@ public class DockerComposeScriptEngine extends AbstractScriptEngine {
         // Catch all exceptions to not sacrifice functionality for logging.
         try {
             org.apache.log4j.PropertyConfigurator.configure(getClass()
-                    .getClassLoader().getResourceAsStream(log4jConfigurationFile));
+                    .getClassLoader().getResourceAsStream(LOG4J_CONFIGURATION_FILE));
         } catch (NullPointerException e) {
-            System.err.println("Log4j configuration file not found: " + log4jConfigurationFile +
+            System.err.println("Log4j configuration file not found: " + LOG4J_CONFIGURATION_FILE +
                     ". Any output for the Docker Compose script engine is disabled.");
         } catch (Exception e) {
-            System.err.println("Log4j initialization failed: " + log4jConfigurationFile +
+            System.err.println("Log4j initialization failed: " + LOG4J_CONFIGURATION_FILE +
                     ". Docker Compose script engine is functional but logging is disabled." +
                     "Stacktrace is: ");
             e.printStackTrace();
@@ -66,34 +70,27 @@ public class DockerComposeScriptEngine extends AbstractScriptEngine {
         Map<String, String> variablesMap = processBuilder.environment();
 
         // Add string bindings as environment variables
-        addBindingToStringMap(context.getBindings(ScriptContext.ENGINE_SCOPE), variablesMap);
+        stringBindingsAdder.addBindingToStringMap(context.getBindings(ScriptContext.ENGINE_SCOPE),
+                variablesMap);
 
         // Add DOCKER_HOST variable to execution environment
         variablesMap.put(DOCKER_HOST_PROPERTY_NAME,
                 DockerComposePropertyLoader.getInstance().getDockerHost());
 
         // Replace variables in configuration file
-        script = replaceVariables(script, variablesMap);
+        String scriptReplacedVariables = variablesReplacer.replaceVariables(script, variablesMap);
 
-        File composeYamlFile = new File(DockerComposeCommandCreator.getYamlFileName());
+        File composeYamlFile = null;
+
         try {
-
-            // Create configuration file
-            if (!composeYamlFile.createNewFile()) {
-                throw new FileAlreadyExistsException("Configuration file already exists: "
-                        + DockerComposeCommandCreator.getYamlFileName());
-            }
-
-            // Force configuration file to disk
-            Writer configFileWriter = new FileWriter(composeYamlFile);
-            configFileWriter.write(script);
-            configFileWriter.close();
+            composeYamlFile = configurationFileWriter.forceFileToDisk(scriptReplacedVariables,
+                    DockerComposeCommandCreator.getYamlFileName());
 
             // Start process
             Process process = processBuilder.start();
 
             // Attach streams
-            ProcessBuilderUtilities.attachStreamsToProcess(process,
+            processBuilderUtilities.attachStreamsToProcess(process,
                     context.getWriter(),
                     context.getErrorWriter(),
                     context.getReader());
@@ -121,15 +118,19 @@ public class DockerComposeScriptEngine extends AbstractScriptEngine {
             // Thread was interrupted somehow, now we need to make sure that
             // the container are stopped even though several interrupts could
             // occur. Therefore start a thread which will remove all container.
-            // TODO: Start thread which will stop and remove container in the background
             log.warn("Execution was terminated and container might need to be terminated manually.");
         } finally {
             // Delete configuration file
-            composeYamlFile.delete();
+            if (composeYamlFile != null) {
+                boolean deleted = composeYamlFile.delete();
+                if (!deleted) {
+                    log.warn("File: "+composeYamlFile.getAbsolutePath()+" was not deleted.");
+                }
+            }
         }
-
         return null;
     }
+
 
     // TODO: Test
     @Override
@@ -150,40 +151,6 @@ public class DockerComposeScriptEngine extends AbstractScriptEngine {
         return eval(stringWriter.toString(), context);
     }
 
-    private String replaceVariables(String script, Map<String, String> variables) {
-        String result = script;
-        // Replace all variables one by one
-        for (Map.Entry<String, String> variable : variables.entrySet()) {
-            if (variable.getValue() != null) {
-                result = result.replace("$" + variable.getKey(), variable.getValue());
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Adds all bindings which are from type @String to the environment map. All other bindings are printed
-     * with toString() to log file.
-     *
-     * @param bindings    Bindings which will be read and added to environment.
-     * @param environment Map<String,String> which will get all Entry<String,String> from the @Bindings
-     */
-    private void addBindingToStringMap(@NotNull Bindings bindings, @NotNull Map<String, String> environment) {
-        for (Map.Entry<String, Object> entry : bindings.entrySet()) {
-            if (entry.getValue() instanceof String) {
-                addEntryToEnvironmentWhichIsAPureString(environment, entry);
-            } else { // Go through maps and add String String values to the environment map.
-                environmentVariablesAdder.AddEntryToEnvironmentOtherThanPureStrings(environment, entry);
-            }
-        }
-    }
-
-    private void addEntryToEnvironmentWhichIsAPureString(@NotNull Map<String, String> environment,
-            Map.Entry<String, Object> entry) {
-        environment.put(entry.getKey(), (String) entry.getValue());
-        log.debug("Added binding: " + entry.getKey() + ":" + entry.getValue().toString());
-    }
-
 
     @Override
     public Bindings createBindings() {
@@ -192,6 +159,6 @@ public class DockerComposeScriptEngine extends AbstractScriptEngine {
 
     @Override
     public ScriptEngineFactory getFactory() {
-        return new DockerComposeScriptEngineFactory();
+        return new DockerComposeScriptEngineFactory(new DockerComposeVersionGetter(processBuilderUtilities));
     }
 }
